Fix 3: Adjusting the structure and the styling of ContactsPage
