+++ front/src/globals/routes.js
@@ -28,5 +28,13 @@ export default [
       shown: false,
       label: "" // generated dynamically
     }
+  },
+  {
+    path: "/contacts",
+    component: ContactsPage,
+    inNavbar: {
+      shown: true,
+      label: "Contacts"
+    }
   }
 ];
