+++ front/src/pages/ContactsPage/components/Info/index.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
-import "./style.css";
-import imgPlaceholder from "../../assests/placeholder.png";
+import "./index.css"
+import imgPlaceholder from "../../../../assests/placeholder.png";
 
 export default function Info(props) {
     return (
