+++ front/src/pages/ContactsPage/index.css
@@ -1,0 +1,36 @@
+.flex-row-2 {
+  display: flex;
+  flex-flow: row wrap;
+}
+
+.flex-row-2 .box {
+  width: 100%;
+  padding: 20px;
+}
+
+.sup-span {
+  color: #dd0000;
+}
+
+.error-message {
+  font-size: 3vw;
+  font-weight: normal;
+  font-style: italic;
+  color: #dd0000;
+}
+
+#Contacts {
+  background-image: url("../../assests/Contact-header.png");
+}
+
+#Contacts header {
+  padding-left: 7%;
+  text-transform: uppercase;
+}
+
+/* Media Queries */
+@media only screen and (min-width: 768px) {
+  .error-message {
+    font-size: 1.5vw;
+  }
+}
