--- front/src/pages/ContactsPage/style.css
@@ -1,148 +1,0 @@
-.flex-row-2 {
-    display: flex;
-    flex-flow: row wrap;
-}
-
-.flex-row-2 .box {
-    width: 100%;
-    padding: 20px;
-}
-
-.sup-span {
-    color: #dd0000;
-}
-
-.error-message {
-    font-size: 3vw;
-    font-weight: normal;
-    font-style: italic;
-    color: #dd0000;
-}
-
-#Contacts header {
-    background-image: url("../../assests/Contact-header.png");
-    background-size: cover;
-    padding-left: 7%;
-    text-transform: uppercase;
-}
-
-#Contacts article#contact-info {
-    padding: 10% 5% 5% 10%;
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-}
-
-#Contacts article#contact-info section#info {
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: space-evenly;
-    width: 100%;
-    padding: 20px;
-}
-
-#Contacts article#contact-info section#info p {
-    background-color: #333c60;
-    border-radius: 5px;
-    width: fit-content;
-    padding: 0 20px;
-    margin: 5% 0;
-}
-
-#Contacts article#contact-info section#map {
-    width: 100%;
-}
-
-#Contacts article#contact-info section#map h2 {
-    width: 100%;
-    text-align: center;
-    margin-bottom: 10%;
-}
-
-#Contacts article#contact-info section#map img {
-    max-width: 100%;
-}
-
-#Contacts article#contact-form {
-    background-color: #111a3e;
-    padding: 7% 10%;
-    font-weight: bolder;
-}
-
-#Contacts article#contact-form .box label,
-#Contacts article#contact-form .box input,
-#Contacts article#contact-form .box textarea {
-    display: block;
-    width: 100%;
-    color: #ffffff;
-}
-
-#Contacts article#contact-form .box input,
-#Contacts article#contact-form .box textarea {
-    padding-left: 20px;
-    font-size: 4vw;
-}
-
-#Contacts article#contact-form .box input:focus,
-#Contacts article#contact-form .box textarea:focus {
-    outline: none;
-}
-
-#Contacts article#contact-form .box input,
-#Contacts article#contact-form .box textarea {
-    background-color: #50556f;
-    border: 3px solid #c7c8d0;
-    border-radius: 10px;
-}
-
-#Contacts article#contact-form textarea {
-    min-height: 200px;
-    color: #ffffff;
-}
-
-#Contacts article#contact-form .flex-row-2 .box {
-    margin-bottom: 3%;
-}
-
-#Contacts article#contact-form input[type="submit"] {
-    display: block;
-    min-height: 5vh;
-    margin: 0 auto;
-    width: 100%;
-    font-weight: bolder;
-    background: #ffffff;
-    color: #3e3e7c;
-    border: none;
-    border-radius: 10px;
-}
-
-/* Media Queries */
-
-@media only screen and (min-width: 768px) {
-    #Contacts article#contact-info {
-        flex-flow: row nowrap;
-    }
-    #Contacts header {
-        min-height: 100vh !important;
-    }
-    #Contacts article#contact-form .box input,
-    #Contacts article#contact-form .box textarea {
-        font-size: 2vw;
-    }
-    #Contacts article#contact-info section#info,
-    #Contacts article#contact-info section#map,
-    .flex-row-2 .box,
-    #Contacts article#contact-form input[type="submit"] {
-        width: 50% !important;
-    }
-
-    .error-message {
-        font-size: 1.5vw;
-    }
-}
-
-@media only screen and (min-width: 1250px) {
-    #Contacts header {
-        min-height: 120vh !important;
-    }
-}
